+++ myalbum-service-provider-backstage/src/main/java/cn/yan_wm/myalbum/service/provider/backstage/service/SysAdminRoleService.java
@@ -17,5 +17,5 @@ public interface SysAdminRoleService {
      * @param roleId
      * @return
      */
-    int insert(Long adminId, Long roleId);
+    int insert(Integer adminId, Integer roleId);
 }
